Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pg8000==1.31.2
alembic==1.13.1
pydantic==2.7.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
openai==1.30.1
Pillow==10.3.0
torch==2.3.0
transformers==4.41.1
sentencepiece==0.2.0
